+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Release date: ??.??.????
 * Support for referring to student articles from the teacher domain (HUB-374)
 * Copy URL button position change (HUB-394)
 * Display only the login/logout main menu item on teacher domain (HUB-397)
+* Display general office hours grouped by language (HUB-362).
+* Do not display news blocks on teacher domain frontpage (HUB-392)
 
 
 ## 1.26